+++ packages/cli/scripts/start.mjs
@@ -453,7 +453,7 @@ async function deployDebugStack(config, cliInfo) {
       config.stage,
       config.region,
       // wrap paths in quotes to handle spaces in user's appPath
-      `"${State.stacksPath(paths.appPath)}"`,
+      `"${paths.appPath}"`,
     ].join(" "),
     output: "cdk.out",
   };