+++ www/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @serverless-stack/docs
 
+## 1.1.1
+
+### Patch Changes
+
+- [#1719](https://github.com/serverless-stack/serverless-stack/pull/1719) [`3610b5da`](https://github.com/serverless-stack/serverless-stack/commit/3610b5da4f57fd35f1fb824701b7bfc4e8ce1a83) Thanks [@fwang](https://github.com/fwang)! - Function: add support for dotnet (.NET 6) runtime
+
+* [#1716](https://github.com/serverless-stack/serverless-stack/pull/1716) [`480ce263`](https://github.com/serverless-stack/serverless-stack/commit/480ce26378962d3ced737624dad4734cd7ad2b53) Thanks [@fwang](https://github.com/fwang)! - Cron: support disabling cron job
+
 ## 1.1.0
 
 ### Minor Changes